requirements.txt
python-telegram-bot==13.15
python-dotenv
supabase
anthropic
urllib3==1.26.16
keybert